+++ src/CryptoCurrency.Core/Exchange/Model/CreateOrder.cs
@@ -21,8 +21,8 @@ namespace CryptoCurrency.Core.Exchange.Model
 
         public OrderSideEnum Side { get; set; }
 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
-        public double Volume { get; set; }
+        public decimal Volume { get; set; }
     }
 }
