+++ src/CryptoCurrency.Core/Market/MarketTick.cs
@@ -11,10 +11,10 @@ namespace CryptoCurrency.Core.Market
 
         public Epoch Epoch { get; set; }
 
-        public double BuyPrice { get; set; }
+        public decimal BuyPrice { get; set; }
 
-        public double SellPrice { get; set; }
+        public decimal SellPrice { get; set; }
 
-        public double LastPrice { get; set; }
+        public decimal LastPrice { get; set; }
     }
 }
